move HostMacroDifference
